bookstore-microservice/order-service/requirements.txt
Django==5.0.7
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
requests==2.32.3
mysqlclient==2.2.0
